+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - lazydev: allow LuaLS to understand vim modules and other things
+- telescope: added manix picker
 
 ## [2.1] - 2026-07-21